rename package ct to common
Its un-go like to not match package name to basedir. Also plays
better with other tools (e.g. gocode and GoToDef...)
