+++ cmd/cguestd/api_test.go
@@ -22,7 +22,7 @@ import (
 )
 
 type APISuite struct {
-	ct.Suite
+	common.Suite
 	Port           uint
 	BeanstalkdCmd  *exec.Cmd
 	BeanstalkdPath string
