+++ cmd/cnetworkd/api_test.go
@@ -14,7 +14,7 @@ import (
 )
 
 type APISuite struct {
-	ct.Suite
+	common.Suite
 	Port      uint
 	APIServer *graceful.Server
 	VLAN      *lochness.VLAN
