+++ cmd/nfirewalld/cmd_test.go
@@ -14,7 +14,7 @@ import (
 )
 
 type CmdSuite struct {
-	ct.Suite
+	common.Suite
 	ConfigPath string
 	BinName    string
 	Hypervisor *lochness.Hypervisor
@@ -25,7 +25,7 @@ type CmdSuite struct {
 func (s *CmdSuite) SetupSuite() {
 	s.Suite.SetupSuite()
 
-	s.Require().NoError(ct.Build(), "failed to build nfirewalld")
+	s.Require().NoError(common.Build(), "failed to build nfirewalld")
 	s.BinName = "nfirewalld"
 }
 
@@ -60,7 +60,7 @@ func (s *CmdSuite) TestCmd() {
 		"-f", s.ConfigPath,
 		"-i", s.Hypervisor.ID,
 	}
-	cmd, err := ct.Exec("./"+s.BinName, args...)
+	cmd, err := common.Exec("./"+s.BinName, args...)
 	s.NoError(err)
 
 	time.Sleep(1 * time.Second)
